% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_get_climate.R
\name{ff_get_climate}
\alias{ff_get_climate}
\title{Download climatic covariates (CHELSA)}
\usage{
ff_get_climate(dem, aoi, variables = "all", overwrite = FALSE, crop = TRUE)
}
\arguments{
\item{dem}{terra::SpatRaster used as spatial template}

\item{aoi}{sf POLYGON or MULTIPOLYGON in EPSG:4326}

\item{variables}{Climatic variables to download. Default "all" or bio01-bio19, kg5, npp}

\item{source}{Climate data source. Default "CHELSA"}

\item{resolution}{Resolution in arc-secondsc("2.5", "5", "10", "0.5"). Default 0.5}

\item{resample_method}{Resampling method. Default "bicubic"}
}
\value{
terra::SpatRaster
}
\description{
Downloads bioclimatic variables from CHELSA and resamples them
to match a DEM template.
}
