% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_get_ogc.R
\name{ff_get_ogc}
\alias{ff_get_ogc}
\title{Generate oblique geographic coordinates (OGC)}
\usage{
ff_get_ogc(dem, n_directions = 6)
}
\arguments{
\item{dem}{A \code{SpatRaster} DEM (projected, planar coordinates).}

\item{n_directions}{Integer. Number of oblique directions to compute.}
}
\value{
A \code{SpatRaster} with oblique coordinate layers.
}
\description{
Computes oblique geographic coordinates based on a DEM following the
approach of Møller et al, 2020 (https://doi.org/10.5194/soil-6-269-2020). These covariates capture broad spatial trends
and are useful as baseline predictors in Digital Soil Mapping.
}
\details{
This function depends on the package \code{OGC}, which is not available on
CRAN. Users must install it manually using:

\preformatted{
devtools::install_bitbucket("abmoeller/ogc/rPackage/OGC")
}
}
