% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_get_rs_indices.R
\name{ff_get_rs_indices}
\alias{ff_get_rs_indices}
\title{Download Sentinel imagery and compute spectral indices}
\usage{
ff_get_rs_indices(
  aoi,
  dem,
  sensor = c("S2", "S1"),
  dates_S1,
  dates_S2,
  s2_composite = "median",
  exclude_domains = c("urban", "snow"),
  verbose = TRUE
)
}
\arguments{
\item{aoi}{An \code{sf} POLYGON or MULTIPOLYGON.}

\item{dem}{A \code{SpatRaster} used as spatial reference.}

\item{sensor}{Character vector. One or both of \code{"S2"}, \code{"S1"}.}

\item{dates_S1, }{dates_S2 Character. Start and end dates range (c(YYYY-MM-DD,YYYY-MM-DD)).}

\item{s2_composite}{Character. Composite function for Sentinel-2.}

\item{exclude_domains}{Character. Index domains to exclude.}

\item{verbose}{Logical.}
}
\value{
A named list of \code{SpatRaster} objects with spectral indices.
}
\description{
Downloads Sentinel-1 and/or Sentinel-2 imagery using the \code{rsi} package
and computes a standardized set of spectral indices for Digital Soil Mapping.
}
