% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_get_terrain.R
\name{ff_get_terrain}
\alias{ff_get_terrain}
\title{Generate terrain covariates using SAGA GIS (Rsagacmd required)}
\usage{
ff_get_terrain(dem, saga_path, tools = "all", cores = 4, verbose = TRUE)
}
\arguments{
\item{dem}{A \code{SpatRaster} DEM (projected, planar coordinates).}

\item{saga_path}{Full path to \code{saga_cmd.exe}.}

\item{tools}{Character. One or more tool groups to run. Options are
\code{"all"}, \code{"basic"}, \code{"hydrologic"},
\code{"channel_network"}, \code{"terrain_classification"},
\code{"morphometric"}, \code{"mrvbf"}.}

\item{cores}{Integer. Number of cores for SAGA.}

\item{verbose}{Logical. Print SAGA messages.}
}
\value{
A \code{SpatRaster} with terrain covariates.
}
\description{
Computes a standardized set of terrain covariates for Digital Soil Mapping
using SAGA GIS. The DEM is filled by default using the Wang & Liu algorithm.
}
